daily_financial_summary/requirements.txt
crewai
litellm
requests
python-dateutil
python-dotenv